src/requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=0.19.0
numpy>=1.21.0
simfin>=0.8.2
matplotlib>=3.5.0
seaborn>=0.11.0